cherp: a few operational and cosmetic changes
Don't want the creation of a project file to require input for the
scratch directory, cherrypick commits list, etc.

Leave that to the new command or the N option from the menu interface.

Changed the presentation of branch selection so the currently selected
branch is highlighted. Added a leading string of instructions.

Changed the show_next_cherrypick function not to output if there is no
current upstream hash saved in the project file.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>