renpat: remove version number when renaming files
Because there is no enforced standard of where to put the version
number, the naming of the files could put them out of order when a
version number is included in the subject line.

This change removes the version number from the subject string before
renaming the file according to the subject string.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>