patchtools: refactor the select next patch logic
The option to choose a patch should not be present unless there are
patch files in the indir.

If there are just mail files that have not been renamed as patches,
there is no way of telling if they contain patches or replies or
something else entirely.

lib/patch-common.source
patchdevelop
patchreview

Signed-off-by: Tony Camuso <tcamuso@redhat.com>