patchtools: cherp: fix hang when no project file
Script would try to read the project file and hang if the file didn't
exist. The fix assures that a project file exists before any atempt
to read it is made.

Was also hanging when trying to read the file of commits, if that
file didn't exist, so don't try to read it if there isn't one.

Allow cherp status to work even if there really isn't much to show.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>