patchtools: cherp: newsession did not set original head
The reason was that set_gitbranch was exiting if the user did not
actually set a specific branch. That is, the user was happy with the
current branch and pressed 'q' to return without making a selection.

The set_gitbranch was incorrectly testing for a returned 'q' in the
variable that was passed by reference, rather than the return value
from the call to loop_range_q().

Signed-off-by: Tony Camuso <tcamuso@redhat.com>