patchtools: moved renpat to lib/patch-common
This move was to facilitate reporting of any discrepency between the
number of patches listed in a set's denominator and the actual number
of files in the set.

Files with identical subjects are only processed once, to deal with
possible redundancies from the mail server.

A bug in the numerator/denominator code was fixed, whereby a '/' in
the subject line of a standalone patch, not part of a series, could
cause the script to crash.

These things were fixed by separating the subject prefix from the
rest of the subject line and processing the prefix for the
numerator/denominator.

An additional check was made to see if the patch is part of a series
by looking for the numerator/denomiator token.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>