patchtools: git prologue exclusion
Have a lot to do yet, but this is a working version.

TODO:
Use only a validpatchlist, and not the patchdirectory to pass to
subsequent scripts. Or maybe give them the ability to derive their
own validpatchlist. That may be the better option, since it allows
the subsequent scripts to remain independent.

Need to set a boolean that says a source (include) file was already
loaded.

	modified:   lib/gitutilities.source
	modified:   lib/patch-common.source
	modified:   patbatcmp
	modified:   patbatcmpmgr
	modified:   patcmp

Signed-off-by: Tony Camuso <tcamuso@redhat.com>