lib/gitutilities.source: WIP: fix null character bug in validate_commit
When a string gets tokenized, the non-alphanumeric characters that
are delimiters are converted to null characters, probably due to
some IFS magic. Ignore the nulls in the tokenized RHEL and upstream
subject lines.

This is working, but there is debug code, commented out, but still
embedded. Not only that, I want to ignore ALL null characters, because
the upstream and RHELL subject lines do not necessarily have the same
non-alphanumeric characters.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>