patchtools: patbatcmp: was not reporting RHEL-only as conflicts
Also was not picking up new config files as RHEL-only, so needed to
change the valid git diff prefix from '--- a/' to '---'

Added a blank line before running the comparison, because the dot
ouputter moves the cursor up one line and deletes the line before
printing any dots. This was deleting the report of the last patch
file on the screen.

Changed the verbose output from "congruency" language to "conflict"
language.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>