patchtools: check_fixes: Add Fixes: to the grep string.
check_fixes()
Looking for just the 12-char hash was yielding false positives.
Adding "Fixes: " to the grep string eliminates them.

select_remrepo()
Return the status from the call to ui::select_from_array(). This
will signal the caller that a 'q' was entered, so not to bother
with any subsequent processing of remote repo/branch selection.

modified:   lib/gitutilities.source
Signed-off-by: Tony Camuso <tcamuso@redhat.com>