patchtools: false positive missing fixes
Was not catching fixes that were included in the series due to a
misspelling of a variable name, commmitsfile instead of commitsfile
in the call to parse_missingfix in function check_fixes.

Also took the opportunity to improve messaging and to only flag the
missing-fixes file only when there really are missing fixes, so a
return value was added to parse_missingfix.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>