patchtools: remove directory arg from call to check_fixes
There was a bug whereby the change to the remote directory for
checking fixes was not reversed, so all new operations tried to
execute from the remote repo directory. That did not work well.

The directory needs to be changed in the calling environment. The
check_fixes() function should not be bothered by that. Having it
cd to the upstream directory confused matters, since there is more
than one place it can exit.

modified:   lib/gitutilities.source
modified:   lib/patch-common.source
Signed-off-by: Tony Camuso <tcamuso@redhat.com>