patchtools: check_fixes wasn't looking in the RHEL branch for the fix
It is possible that the fix was merged into RHEL earlier than the
current patch series being reviewed. Before having an external
upstream repo directory, that was covered. Now, the code must be
in the RHEL repo directory to test for that.

Also use the time stamp of the commit being sought to shorten the
time it takes to look for it.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>