lib/patch-common: exec_patcmp: fix quiet bug
shellcheck complains if you don't use quotes, but there are some
places where they should not be used. This is one example where
an option variable passed to an external script can be empty.
With quotes around it, it counts as a command line argument, even
though it's just a space, and that confounds the options parser.
So, no quotes around options variables.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>