patchtools: patbatcmp: bug fixes
For v12, I thought it would be a good idea to print the number of
files being compared into the mismatch log. That was unnecessary
and caused a problem by ofsetting the verbose level in the file.
The print of the number of files was removed.

shellcheck changes for v12 inadvertently bugged the logic that
detects and reports "no upstream patch", aka RHEL-only, so that
was fixed.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>