patchtools: make the approve menu work like a second main menu
This was an extensive change requiring some refactoring. The challenge
was to allow the Approve menu to be called from more than one place
without recursion.

The Approve menu was not originally intended to operate as such, but
the convenience was too good to forego. First attempt a few revisions
ago only sort-of worked. This fixes the bugginess of that first attempt.

lib/patch-common.source
lib/patch-mrutilities.source
patchreview

Signed-off-by: Tony Camuso <tcamuso@redhat.com>