patchtools: patch-mrutilities: add current mr to review history
Instead of specifying the current mr, let it be the default for
review history. Make specifying a different mr for review history
a different menu entry.

Fixed an offset bug in the update_history function.

Made the mr field width in the history relative to the number of
digits in the mr id.

Reworked some of the comments and added some new ones for some of
the more elusive or obscure scripting.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>