patchtools: cherp: some cleanup from recent changes
Don't need to store last good commit in cherp.conf, because it's
simply the current HEAD, easily obtainable.

Get the current subject prefix from calling git config to simplify
code.

Only set dnorighead in one place, and that's init_new_session. The
only exception is that if it's NULL when cherp is invoked, and the
it's set to HEAD.

Since we're using git cherry-pick again, there's no need to reset
hard to the last good commit to synchronize the index in function
init_current_session. git cherry-pick does that for us.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>