patchtools: mkbackportlog: improvements in menu choices
Added the option to add paths to the existing search path.
Changed the requirement for a downstream commit to choolse
"One of the following", upstream commit or downstream commit.
Had to deal with getting the search criteria back from getuser,
when I expected to get the expansion of the the command.
Moved the checking of user inputs to test_reqparms.
Some comment cleanup.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>