patchtools: cherp: finish the reset functionality
A lot of ancillary changes. Comment updates, etc.
But the reset_cherp and get_newhead functions were the real focus
of this commit.

Functionality is complete, so users can reset the state of their
development branch to anywhere in its history.

TODO: Would like to be able to add the ability to go back from
earlier commits to later commits, but the preponderance of the
desired functionality is there.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>