patchtools: cherp: two fixes and a wcherp accomodation
Fix 1. In get_lastuplog, don't tail a file that may not exist. That
       causes hangs.
Fix 2. In read session, using 'git remote show origin' to get the
       origin branch name is problematic when there is no connection
       to the remote repo. That hangs the script.
       The solution is to extract the name for the remote branch from
       the .git/config file.

The accommodation was for wcherp. There is no deterministic way to get
the terminal background color in the tmux environment. Ergo, if the
terminal backtground color hasn't been set by using cherp on its own,
then the user must be queried for it. The user's choice is written
into the cherp.conf file, so it is sticky across invocations.

The lib/cherp.conf template file was updated to include the background
color field, tbackground.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>