patchtools: getup: WIP: do away with git cherry-pick
Major refactoring due to replacing git cherry-pick with git am -3.
The reason was to avoid having to replicate the upstream repo in
the user's local downstream directory as a remote.

With git am -3, the same kind of merge occurs, but without requiring
the upstream as a remote in the current directory.

Also took the liberty to rename some variables and functions more
intuitively and to reflect the removal of the cherry-pick funcionality.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>