patchtools: gorp: cleanup after functional test
- Functional testing revealed some flaws introduced by using faster
methods like awk and sed, instead of parsing strings in whole files.
- Had to juggle some initialization to be sure the global strings were
initialized for both interactive menu and command line.
- Introduced mark_out function to remove upstream backports that have
already been committed. This can happen when the Original Head is
updated with a git pull and some of the commits in the backport
file are in the update.
- Used git apply --reverse --check  < <(git show "$uphash") in the
backport function to test whether a commit has already been
backported.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>